Use Picard-like arguments for Read Groups
As we are using Picard/GATK as our starting point, we would like to
maintain as much similarity as possible regarding the command line to
allow consistent pipelines combining ReadTools with the other toolkits.

Thus, in this commit we deprecate the ReadGroupLegacyArgumentCollection
in favor of a new ReadGroupArgumentCollection, which in addition is
fully tested.

Only a small change in the platform value type differs between Picard and RT.