Fix bug in SplitGATKWriter
If SplitGATKWriter is splitting by a list which contains similar
mappings for different records an onDemand=false, the initiliazation
blows up because it is trying to create a new reader for the key.

This bug fix also includes a test that is failing before and fixed by
the patch, but also test the write/read of one of the split outputs.