Solve issue of different queryname sort orders
Because 'queryname' is not defined in the specs and we need only to keep
the read pairs together, this patch does not assign 'queryname' unless
it is already present in the header. See #87 for more information.

In addition, it logs a warning if the header is sorted in that way to
point out that we follow the Picard sort order and this is assumed.
Other warnings are updated to reflect our expectations regarding
ordering of pair-end data.