Add hack for GATKReadFilterPluginDescriptor
The current implementation of GATKReadFilterPluginDescriptor in GATK
have some limitations for our use case. Some of the parts of this hack
will be included in GATK but they require some review by the team, but
we need this into ReadTools soon.

Thus, I added the plugin descriptor "as is" from GATK, with the changes
included in the PRs. In addition, I included some tests to make sure
that is working properly. Finally, I added the new method
getDefaultInstances from the new interface in Barclay.

Note: this classes will be removed once GATK accept them.