Add new trimming arguments and refactor TrimmerPluginDescriptor
For the new set of tools, the arguments for trimming are going to be
more consistents and rely on the plugin descriptor for trimmers and
filters. This commit includes a new trimming argument collection
(renaming the previous one to point to legacy code) and refactor the
TrimmerPluginDescriptor as necessary for its integration with the
contract of plugin descriptors.

Some changes are included to facilitate the transition to the next
version of Barclay.