Disable/Enable read end trimming within TrimminFunction
The previous design of the TrimmingFunction and integration pipeline
delegated the enabling/disabling of the trimmer to
ApplyTrimResultReadTransformer, causing problems when checking if a read
is completely trimmed or not within the TrimmingFunction. This commit
refactor all the new trimming engine to include the 5/3 prime disabling
directly into each trimmer, allowing several simplifications:

- Combination of arguments in TrimmerPluginDescriptor
- Simplification of ApplyTrimResultReadTransformer
- Keeping trimming logic inside TrimmingFunction
- Simplification of TrimAndFilterPipeline

Note: it includes a bug fix in ApplyTrimResultReadTransformer when the
tags are removed.