Fix minor bug in Mott's algorithm
If a read is completely trim, it returns an array which does not reflect
in both ends that it was trimmed. This is bad for the purpose of
disabling one or the other end trimming, so now it reflects the real
state.

This change may introduce more complexity if used with
Arrays.copyOfRange, because the new behaviour now blows up with an
IllegalArgumentException.