Update GATK read filter plugin descriptor hack (#145)
Based on the GATK pull request to fix some issues with
GATKReadFilterPluginDescriptor, this commit includes the current changes
to check for suitability/compatibility of the new implementation for our
purposes.

The class should live still in ReadTools, because there are some minor
changes with respect to the GATK version. This is due to some problems in
the Barclay contract for the plugin that may be changed in the near
future.