Integrate Distmap support for output (#152)
Distmap is a software for mapping in the Hadoop cluster, and this commit includes a tool for convert the reads to the distmap format to be used afterwards by the program. If the output path is HDFS, this allow to upload the reads directly to Hadoop. This improve the behaviour of making a local copy and then upload to HDFS, at the cost of network connection.

Concretely, this commit includes:

- Utility class for encoding/decoding reads in Distmap format
- GATKWriter for output reads in this format
- Tool for upload reads for Distmap (supporting HDFS block-size setting)
- Allow bzip2 compression through Hadoop codec (for both local/hdfs files)