Concatenate barcode qualities by space (#221)
This is a no backwards-compatible change that is related with the 
SAM-specifications and barcode tags (BC/QT). Concretely, there is 
an ongoing discussion about this at the SAM-specs repository: 
barcode qualities cannot be separated by hyphens because it is 
a valid quality string.

This commit implements the current solution discussed in the
SAM-specs repository, which is to concatenate QT with spaces.
BC tag is still concatenating sequences with hyphens.