Add new FAQ entry and reformat FAQ page template
The previous FAQ template was incorrect when adding a new FAQ entry,
so now the questions are organized inside a YML file and the HTML
iterate over the entries to add all the questions.

Thus, anytime a new question should be included a new entry to the
YML should be added to include in the final output.