Fix broken javadoc (#281)
Due to an error in the Java's code for javadoc generation, 
an annotation method could not be accessed if it has a 
default value. This commit fixes the usage of this broken 
feature in our javadoc.

It also fixes the travis script for testing several 
commands, which before this commit was only failing 
if the last command does.