Refactor FASTQ common test data (#289)
To add more visibility to the type of formats tested in ReadTools, this
commit adds a new folder in the root directory for test data that may be
used to test the program outside the test suite, with a README including
which kind of data is contained. This also includes some redundant file
cleaning.

The new common test data could be accessed in tests through
TestResourcesUtils.