Support CRAM input with java.nio.Path reference (#368)
This commit adds support for Path reference in DistmapPartDownloader and
ReadReaderFactory (which is supposed to be used in every tool/engine
except the distmap downloader).

The support implemented here is only for reading, not for writing;
nevertheless, the interfaces for writing were changed to take a Path
instead to consistently handling the input.

WARNING: java.nio.Path in writers is unsupported due to HTSJDK
limitations. This may change in the future.