Upgrade GATK to 4.beta.6 (#378)
Upgrading to GATK 4.0.0.0 is not possible yet, because there
are some incompatibilities with respect to the arguments that
are in use (concretely, regarding ReadFilters and
CommandLineProgram).

Until we can fix safely that issues, this is a minor version bump
with almost no changes. We removed a dependency that will
be removed in GATK 4.0.0.0 in favor of a Picard class,
concretely in BarcodeDictionaryFactory.