Handle compression for HDFS with Hadoop compression codecs
Before this commit only bzip2 was supported for any output file,
automatically detected with the .bz2 extension. Nevertheless, this does
not allow to use the potential of the Hadoop service for compression
codecs, which allows downstream users to pass a custom classpath for
other compression algorithms not bundled in ReadTools.

This changes include:

* Extract a HadoopUtils class with methods to interact with HDFS and
  Hadoop
* Change the block-size argument to long
* Remove usage of hadoop-library in ReadWriterFactory to depend only in
  HadoopUtils
* Change bzip2 support for non-HDFS Path to apache-compress
* Tests for compression with HDFS codecs

This is also a step torward a two-jar distribution, the current uber-jar
and one excluding Hadoop/Spark dependencies. The later might be usefull
for work in conjuction with Hadoop/Spark classpaths specified by the
user (or using hdfs).