Fix license in MANIFEST failing in OSX
In MacOS, files and folders cannot have the same name and because the
File System is case-sensitive, uncompressing the packaged jar fails due
to a collision in names between the LICENSE file and the license/
folder.

This commit fixes the error by using the license transformer from the
gradle shadow plugin (ApacheLicenseResourceTransformer).