Upgrade to GATK 4.0.2.1-ReadTools and HTSJDK 1.14.3 (#442)
This commit upgrades GATK to a released version, although depends
on a patched repository. This is important to allow backwards-compatible
command line (ReadTools version 1). For version 2.0.0, this dependency
will be removed to remove backwards compatibility in the command line.

In addition, it updates htsjdk to version 1.14.3 and introduces the following
minor changes:

- Program groups moved to Picard
- Changes in barclay parser for plugins (TrimmerPluginDescriptor now
  follow the contract again)
- Print correct version in startup message (through MANIFEST inclusion)
- Logs the usage of patched GATK to inform the user