Add ComputeProperStatByWindow experimental tool
Implements a tool to count the number of reads over window with tags
w.r.t. a threshold, with optional counts for reads with
indels/softclips. This tool is experimental because it still requires
integration with a barclay plugin to allow compute the statistics with
user-specified StatFunctions (not mandatory int-tag-counter) and with
more statistic implementations.

In addition, this commit introduces the following changes:

- ReadTools extension of GATK's ReadWalker (RTReadWalker)
- New program group for mapped reads
- Temporary argument collection for ComputeProperStatByWindow (instead
  of barclay plugin)