Add DownloadDistmapResult check for _SUCCESS on startup
This adds a check to the _SUCCESS file on startup to avoid downloading
truncated data due to non-termination of the MapReduce job. This check
can be disable with an advance argument to allow the user to download
data for debugging, but it is not recommended.

In addition, this commit reformats the code a bit without logic impact
and with improved error messages for the download of only some parts
(debug argument).