Implement experimental GemMappabilityToBed tool
New experimental tool to perform conversion between GEM-mappability
and BED-graph format. This tool includes also an scoring method switch,
to convert the range of number of mappings (n) to an unique value. Current
scoring methods are minimum/maximum/middle. In the future, this will be
extended to score also the "mappability" sensu stricto (1/n).

This commit also adds a new program group for mappability-related tools.