Force locale for base test class
RTBaseTest extends BaseTest, but none of them set the locale to US-EN as
the Main class does. Thus, test dependent on formatting and other
locale-related classes that does not call Main might fail in other
systems.

This is a temporary fix until we backport to GATK the call to force the
JVM in tests to use the same locale.