Barcode file with empty header throws user error (#520)
This improves the help message for a barcode file were, by
an user error, there are more than one tab between columns.

This commit checks for null comlumn labels, which indicates this
error. It does not check for null elements on the table, as it is
supported to set to null some parts of the read groups.